[Common] Move common code to crate
