+++ crates/common/src/lib.rs
@@ -1,7 +1,3 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
+pub mod capability;
+pub mod seccomp;
+pub mod unix_stream;
