--- src/lib.rs
@@ -1,13 +1,0 @@
-//! This is the main library interface for this project
-#![deny(missing_docs)]
-
-mod capability;
-mod ffi;
-mod oci;
-mod sandbox;
-mod seccomp;
-mod storage;
-mod unix_stream;
-
-pub mod kubernetes;
-pub mod network;
